requirements.txt
#psycopg[binary]==3.1.10
django==5.0.1
python-decouple==3.8
pandas==2.2.0
mysqlclient==2.2.4